solutions/server/src/requirements.txt
aioredis==1.3.1
altgraph==0.17
asgiref==3.2.10
astroid==2.4.2
async-timeout==3.0.1
attrs==20.2.0
auto-py-to-exe==2.7.8
autobahn==20.7.1
Automat==20.2.0
autopep8==1.5.3
bottle==0.12.18
bottle-websocket==0.2.9
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
click==7.1.2
colorama==0.4.3
constantly==15.1.0
cryptography==3.1.1
daphne==2.5.0
defusedxml==0.6.0
dj-database-url==0.5.0
dj-rest-auth==1.1.2
Django==3.0.8
django-allauth==0.42.0
django-cors-headers==3.5.0
django-heroku==0.3.1
djangorestframework==3.12.1
Eel==0.12.4
Flask==1.1.2
future==0.18.2
gevent==20.9.0
gevent-websocket==0.10.1
greenlet==0.4.17
gunicorn==20.0.4
hiredis==1.1.0
hyperlink==20.0.1
idna==2.10
incremental==17.5.0
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.11.2
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
msgpack==1.0.0
msgpack-python==0.5.6
oauthlib==3.1.0
pefile==2019.4.18
Pillow==7.2.0
psycopg2==2.8.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.6.0
pycparser==2.20
PyHamcrest==2.0.2
pyinstaller==4.0
pyinstaller-hooks-contrib==2020.8
pylint==2.5.3
pyOpenSSL==19.1.0
python-dotenv==0.14.0
python3-openid==3.2.0
pytz==2020.1
pywin32-ctypes==0.2.0
redis==2.10.6
requests==2.24.0
requests-oauthlib==1.3.0
service-identity==18.1.0
six==1.15.0
sqlparse==0.3.1
toml==0.10.1
Twisted==20.3.0
txaio==20.4.1
urllib3==1.25.10
Werkzeug==1.0.1
whichcraft==0.6.1
whitenoise==5.2.0
wrapt==1.12.1
zope.event==4.5.0
zope.interface==5.1.0
